Fix run_#_annualSummary.txt file writing.
e.summary was still a data.frame and was causing the warning messages
because NAs were being generated. In some runs e.summary was also
inheriting a list object from d.summary and t.summary which was
generating an error and crashing the model. Did two test runs and model
did not crash or generate warning messages.