Corrected Model
A number of corrections to the model that address errors from test runs.
Corrections are to both the treatment and wildfire loops. Corrections to
the treatment loop address problems with objects that track area treated
for each management option. Initially the code modified expected area if
there was no available areas for treatment but this caused the model to
crash under certain circumstances. In the current version a seperate
object tracks untreated area, adds it to treated area and compares the
sum with expected area to determine when the loop ends. Additionally, I
have changed the way the treatment loop cycles within loop 3. There was
a coding error at the beginning of this loop that would force an early
exit to the next treatment if expected treatment area could not be
mapped in block one. In the wildfire there were errors in the section of
loop 9 leading up to loops 10 and 11 that would cause the model to crash
under certain conditions. These errors were related to how seed cells
were used in loop 10 vs. 11 and how probability of ignition is
calculated.