Fix to loop 4
Added code to loop 4 that will assess probability of ignition for each
cell being considered by the fire spread model during prescribed fires.
This component is not activated for silvicultural treatments. For these
treatments probability of spread is binary and based on the receptivity
of the fuelbed to treatment and if the cell has already been treated
during this current model year.