Fixed error caused by temporary stand number handling in wildfire loop (8)
Temporary stand numbers are stored in objects that begin with tesn... In
the wildfire loop where there are two sub-models (wildfire; loop 10 and
block and burn; loop 11) a single fire can be passed back and forth
between the two. Depending on the history of fire growth tesn's must be
handled differently and must be reset to NULL at the top of loop 9. This
commit includes changes and corrections to code in loops 9 and 10
including a reset at the top of loop 9, increased specificity of how
fire growth is evaluated for tesn assignments at the beginning of loop
10, addition of tesn to if else evaluation function at the top of 9.7.2.

Several other minor errors were also corrected including the removal of
a function that ordered loopB by old stand numbers and addition of a new
fuelbeds column  to the loopB object, both at the end of the prescribed
fire loop (loop 2),  and the reset of the initial.new.cells object in
11.2.1.

Not an error, but I also removed the system.function that wrapped around
loop 1.