Fix small error in lookup table and script
Update fuelbed lookup table by removing instance of fuelbed number
4015301 in the post columns. Changed to 4015401 because fuelbed 4015301
does not exist in the master fuelbed list. Used R script to check for
other instances of post fuelbed not in the master list (column 1) and
could not find any.

Changed file path for fuelbed map output from a folder within output
path specified at the beginning of the script to just the output path.
This would have either generated a new folder in the S3 bucket or caused
the file to not write and perhaps crash the model.