Fix error generated by no treatment/no fire run
Trial runs with almost no disturbance revealed another error. As with
others this had to do with objects that track time-since-last-treatment.
Specifically the objects that store transition fuelbeds for changes in
mFRI for new post-treatment fuelbeds (new_lower, new_upper_1, and
new_upper_2) are output as an empty list when there are no new
post-treatment fuelbeds. This list() object is passed on to the
mfri_lower and mfri_upper .List objects which pass it on to the
Fuelbed.List which in turn crashes FDM. Fixed this by using a if()
function to skip over these objects and any downstream objects.

During the process of diagnosing this error I also noticed another error
that was equally as serious. New fuelbeds produced by crown fire were
not being assigned to the correct stands. This was fixed by moving up
the function that orders loopE_crownFire  by new stands above the point
in the code where new fuelbeds and new age (also ordered by new stands )
are added to loopE_crownFire.

I also ammended one of the break functions to kill the model if
Fuelbed.List turns into a list object.