Fix error in Burn Block lookup table
This table regulates which management units (i.e., burn blocks) can be
selected for treatment with thinning, herbicide, or prescribed fire. For
the NotCCA column (areas outside of intensive management where Eglin
conducts treatments when prior obligations in more intensively managed
units are complete) you forgot to exclude units that include the No Data
area of the map, the buffer zone, and unmanaged areas of Eglin. None of
these can be treated. The first two for obvious reasons, the last one
because this is 76,266 acre unit that includes many of the built up
areas of the base and a treatment would cover the entire area. Ideally,
the model would be set up to establish a new burn block in areas
affected by natural disturbance such as a wildfire. Until then this area
cannot be treated.

ID'ed this mistake when run 193 triggered a 68,000 acre prescribed fire
in the buffer zone.