THE BUG: Extraneous slash in path to bash
The bug that blocked work on this project for over two weeks turned out to be an extraneous slash on the path to bash. /bin/bash/ points to a directory, which doesn't exist and can not execute a script. /bin/bash is what we wanted, on the other hand.

This bug caused several weeks of delay and hours of manpower. We ended up creating an entirely new AMI in Linux, redownloading R, redownloading all of the R libraries, etc. This is partially the fault of the testing process; when I encountered problems with the script, I copied it line-by-line in the terminal instead of scp'ing over and trying to execute from scratch. While relieved this is solved, a mistake of this nature would cause a more religious person to question that religion.

Lessons learned: Do not omit even seemingly guaranteed things such as script's interpreter directive in testing. A raw /bin/bash instead of using slash variable may have (or may not have) been easier to read and interpret, so use clearer scripts when possible.